Project Complete!
Project of galaxy generator has been entirely completed, for the sake of not giving the user the option to lag themselves out, I've taken out the panel in the comments, but this can be re-enabled by the user by simply getting rid of the comments.

Project is deployed on vercel.
Project will soon have custom domain set up via AWS.